governance/requirements.txt
# governance/requirements.txt
kafka-python==2.0.2
minio==7.2.3
